pyproject.toml
[tool.poetry]
name = "chessmate"
version = "1.0.0"
description = "Automated chess board"
authors = ["Group 5"]

[tool.poetry.dependencies]
python = "^3.11"
flask = ">=2.2.5"
authlib = ">=1.3.1"
python-dotenv = ">=1.0.1"
requests = ">=2.31.0"
python-lichess = ">=0.10"
berserk = ">=0.13.2"
opencv-python = "^4.10.0.82"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"